Add counter_cache to Node and ppm attributes to Datum

+++ db/migrate/20161122022043_create_data.rb
@@ -2,8 +2,8 @@ class CreateData < ActiveRecord::Migration[5.0]
   def change
     create_table :data do |t|
       t.references :node, foreign_key: true
-      t.float :temperature
-      t.float :luminosity
+      t.integer :temperature
+      t.integer :luminosity
 
       t.timestamps
     end
